\name{mvtnorm-internal}
\alias{mvt}
\alias{checkmvArgs}
\title{Internal mvtnorm function}
\description{
 Internal mvtnorm functions.
}
\usage{
mvt(lower, upper, df, corr, delta, maxpts = 25000, abseps = 0.001,
           releps = 0)
checkmvArgs(lower, upper, mean, corr, sigma)
}
\details{
  This functions are not to be called by the user.
}
\keyword{internal}
